# openapi.json
{
  "components": {
    "schemas": {
...
      "FeedbackCreate": {
        "properties": {
          "author_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Author Name",
            "type": "string"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "author_name",
          "message"
        ],
        "title": "FeedbackCreate",
        "type": "object"
      },
      "FeedbackResponse": {
        "properties": {
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author_name",
          "message",
          "created_at"
        ],
        "title": "FeedbackResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchupOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_championship": {
            "title": "Is Championship",
            "type": "boolean"
          },
          "is_playoff": {
            "title": "Is Playoff",
            "type": "boolean"
          },
          "league_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League Id"
          },
          "margin": {
            "title": "Margin",
            "type": "number"
          },
          "season_year": {
            "title": "Season Year",
            "type": "integer"
          },
          "team1_manager_id": {
            "title": "Team1 Manager Id",
            "type": "integer"
          },
          "team1_manager_name": {
            "title": "Team1 Manager Name",
            "type": "string"
          },
          "team1_points": {
            "title": "Team1 Points",
            "type": "number"
          },
          "team1_projected": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team1 Projected"
          },
          "team1_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team1 Team Name"
          },
          "team1_yahoo_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team1 Yahoo Id"
          },
          "team2_manager_id": {
            "title": "Team2 Manager Id",
            "type": "integer"
          },
          "team2_manager_name": {
            "title": "Team2 Manager Name",
            "type": "string"
          },
          "team2_points": {
            "title": "Team2 Points",
            "type": "number"
          },
          "team2_projected": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team2 Projected"
          },
          "team2_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team2 Team Name"
          },
          "team2_yahoo_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team2 Yahoo Id"
          },
          "week": {
            "title": "Week",
            "type": "integer"
          },
          "winner_manager_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Manager Id"
          }
        },
        "required": [
          "id",
          "season_year",
          "week",
          "team1_manager_id",
          "team1_manager_name",
          "team1_team_name",
          "team1_points",
          "team1_projected",
          "team2_manager_id",
          "team2_manager_name",
          "team2_team_name",
          "team2_points",
          "team2_projected",
          "winner_manager_id",
          "is_playoff",
          "is_championship",
          "margin",
          "league_id",
          "team1_yahoo_id",
          "team2_yahoo_id"
        ],
        "title": "MatchupOut",
        "type": "object"
      },
...
      "SeasonDetail": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "league_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League Name"
          },
          "num_playoff_teams": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Playoff Teams"
          },
          "num_regular_season_weeks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Regular Season Weeks"
          },
          "num_teams": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Teams"
          },
          "standings": {
            "items": {
              "$ref": "#/components/schemas/StandingsRow"
            },
            "title": "Standings",
            "type": "array"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "year",
          "league_name",
          "num_teams",
          "num_playoff_teams",
          "num_regular_season_weeks",
          "standings"
        ],
        "title": "SeasonDetail",
        "type": "object"
      },
...
      "StandingsRow": {
        "properties": {
          "final_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Rank"
          },
          "is_champion": {
            "title": "Is Champion",
            "type": "boolean"
          },
          "losses": {
            "title": "Losses",
            "type": "integer"
          },
          "made_playoffs": {
            "title": "Made Playoffs",
            "type": "boolean"
          },
          "manager_id": {
            "title": "Manager Id",
            "type": "integer"
          },
          "manager_name": {
            "title": "Manager Name",
            "type": "string"
          },
          "playoff_finish": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playoff Finish"
          },
          "points_against": {
            "title": "Points Against",
            "type": "number"
          },
          "points_for": {
            "title": "Points For",
            "type": "number"
          },
          "team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          },
          "ties": {
            "title": "Ties",
            "type": "integer"
          },
          "wins": {
            "title": "Wins",
            "type": "integer"
          }
        },
        "required": [
          "final_rank",
          "team_name",
          "manager_id",
          "manager_name",
          "wins",
          "losses",
          "ties",
          "points_for",
          "points_against",
          "made_playoffs",
          "is_champion",
          "playoff_finish"
        ],
        "title": "StandingsRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fantasy Football Dashboard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/feedback": {
      "get": {
        "operationId": "list_feedback_api_feedback_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedbackResponse"
                  },
                  "title": "Response List Feedback Api Feedback Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feedback",
        "tags": [
          "feedback"
        ]
      },
      "post": {
        "operationId": "create_feedback_api_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/api/seasons/{year}": {
      "get": {
        "operationId": "get_season_api_seasons__year__get",
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Season",
        "tags": [
          "seasons"
        ]
      }
    },
    "/api/seasons/{year}/matchups": {
      "get": {
        "operationId": "get_season_matchups_api_seasons__year__matchups_get",
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "week",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchupOut"
                  },
                  "title": "Response Get Season Matchups Api Seasons  Year  Matchups Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Season Matchups",
        "tags": [
          "seasons"
        ]
      }
    },
...
  }
}